% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixest.R
\name{as_tibble.fixest}
\alias{as_tibble.fixest}
\title{Convert a \code{fixest} object to a nested tibble}
\usage{
\method{as_tibble}{fixest}(x)
}
\arguments{
\item{x}{A \code{fixest} object}
}
\value{
A tibble with one row and two columns, \code{id = 1L}, and \code{model = list(x)}.
}
\description{
Convert a \code{fixest} object to a nested tibble, allowing easy
tidying, along with multiple estimation models nested using
\code{as_tibble.fixest_multi()}.
}
