% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixest.R
\name{as_tibble.fixest_multi}
\alias{as_tibble.fixest_multi}
\title{Convert a \code{fixest_multi} object to a nested tibble}
\usage{
\method{as_tibble}{fixest_multi}(x)
}
\arguments{
\item{x}{A \code{fixest_multi} object}
}
\value{
A tibble with model parameters and model IDs as columns and a list
column \code{model} with all models from \code{x}. Note that the tibble will
not contain columns for parameters which do not vary within \code{x}
(see examples).
}
\description{
Convert a \code{fixest_multi} object to a nested tibble, allowing easy
tidying of the models.
}
\details{
Only an \code{as_tibble} method is provided, no \code{as.data.frame} because
printing lists in a data.frame is often problematic.
\code{\link[fixest:models]{fixest::models()}}, which is used internally, always returns a
data.frame.
}
\examples{
# Results only in an `lhs` column
m1 <- fixest::feols(c(mpg, qsec) ~ wt, data = mtcars)
tibble::as_tibble(m1)

# Results only in an `rhs` column
m2 <- fixest::feols(mpg ~ csw(wt, disp), data = mtcars)
tibble::as_tibble(m2)

}
