% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asmisc.R
\docType{package}
\name{asmisc-package}
\alias{asmisc}
\alias{asmisc-package}
\title{asmisc: Miscellaneous Utility Functions}
\description{
A small collection if utility functions, mostly for personal use.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/svraka/asmisc}
  \item Report bugs at \url{https://gitlab.com/svraka/asmisc/issues}
}

}
\author{
\strong{Maintainer}: András Svraka \email{svraka.andras@gmail.com}

}
