% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{callback_write_parquet}
\alias{callback_write_parquet}
\title{Callback function to write Parquet partition}
\usage{
callback_write_parquet(chunk_paths, chunk_size, processing_function = NULL)
}
\arguments{
\item{chunk_size}{The number of rows to include in each chunk}

\item{processing_function}{A function that takes each chunk and
does arbitrary data processing on it before writing the resulting
data frame into its Parquet partition.}
}
\value{
A function to be used in
  \code{\link{read_delim_chunked_to_dataset}}.
}
\description{
A function factory that creates a callback function for
\code{\link[readr]{read_delim_chunked}}.
}
\seealso{
\url{https://stackoverflow.com/a/49241426}
}
\keyword{internal}
