% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chr_values}
\alias{chr_values}
\title{Helper function to list top values of a character vector}
\usage{
chr_values(x)
}
\arguments{
\item{x}{A character vector}
}
\description{
List the first 10 unique values of a character vector.
}
\keyword{internal}
