% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_names}
\alias{clean_names}
\title{Clean column names}
\usage{
clean_names(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
The data frame with cleaned column names.
}
\description{
Clean column names in a data frame.
}
\details{
This function makes all column names lowercase, strips diacritics
  and punctuation marks, and replaces whitespace with underscores.
}
