% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{Create a codebook from a data frame}
\usage{
codebook(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A tibble from a custom wide format \code{\link[skimr]{skim}} output.  Unlike
\code{\link[skimr]{skim}}, histograms are not generated but there are
additional columns:

\describe{
  \item{\code{is_whole}}{Logical, whether a numeric variable has only whole
    values.}
  \item{\code{\link[sjmisc:is_num_fac]{is_num_chr}}}{Logical, whether a
    character column has only numeric values.}
  \item{\code{chr_values}}{A list column with the first 10 unique character
    values.}
  \item{\code{lgl_counts}}{A list column with counts of logical levels.}
  \item{\code{fct_counts}}{A list column with counts of factor levels.}

}
}
\description{
Create a codebook of a data frame using \code{\link[skimr]{skim}}.
}
\examples{
codebook(mtcars)
}
\seealso{
\code{\link{codebook_chunked}} for files that are too big to be read
  into memory at once.
}
