% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook_chunked}
\alias{codebook_chunked}
\title{Create codebook by chunks}
\usage{
codebook_chunked(file, input_delim, chunk_size, process_function = NULL,
  simplify = TRUE, ...)
}
\arguments{
\item{file}{Path to a file.  Passed to
\code{\link[readr]{read_delim_chunked}}, see further details there.}

\item{input_delim}{Single character used to separate fields within a record.}

\item{chunk_size}{The number of rows to include in each chunk.  See
\code{\link[readr]{read_delim_chunked}}}

\item{process_function}{A function that takes a data frame as its input and
processes this data frame.  \code{process_function} is applied to each
chunk before creating the codebook for that chunk.  There are no
restrictions what this function can be.  When left as \code{NULL}, no
processing is done.}

\item{simplify}{Simplify the results.}

\item{...}{Additional arguments passed to
\code{\link[readr]{read_delim_chunked}}}
}
\value{
If \code{simplify} is \code{TRUE}, a list with the same structure as
  \code{\link{codebook}} but each tibble has an additional column named
  \code{chunk} with the number of the chunk it is describing and there are
  additional rows for each variable in each chunk.  Otherwise, a list of
  lists returned by \code{\link{codebook}}.
}
\description{
\code{codebook_chunked} uses \code{codebook} to create codebooks for chunks
of a file, as read by \code{\link[readr]{read_delim_chunked}}.
}
\examples{
codebook_chunked(
  file = readr::readr_example("mtcars.csv"),
  input_delim = ",",
  chunk_size = 10
)
}
