% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_sizes}
\alias{df_sizes}
\title{List data frames and their sizes}
\usage{
df_sizes()
}
\value{
A tibble, arranged in decreasing order of size.
}
\description{
List data frames in global environment and their sizes in megabytes.
}
