% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicates}
\alias{duplicates}
\title{Inspect duplicate rows}
\usage{
duplicates(x, by_list, check_all = FALSE)
}
\arguments{
\item{x}{A data frame, or a data frame extension, like a
\code{\link[tibble:tibble]{tibble::tibble()}}, or a \code{\link[data.table:data.table]{data.table::data.table()}}.}

\item{by_list}{A list, where each element is a character vector
with columns names from \code{x}, indicating which combinations of
columns from \code{x} to use for duplicate checks.}

\item{check_all}{If \code{TRUE}, include number of duplicates by all
columns in \code{x}.}
}
\value{
A \code{data.frame} with columns \code{by} (list), \code{N_unique} and
\code{N_duplicated} (integers), where each row corresponds to an element
of \code{by_list}. If \code{check_all == TRUE}, we add a row to the bottom,
where the value of \code{by} is \code{NULL}.
}
\description{
Inspect duplicate rows in a data frame by sets of columns.
}
\details{
If \code{x} is a \code{data.table}, we use \pkg{data.table}'s optimized
\code{data.table::uniqueN()}, otherwise we use \code{dplyr::distinct()} to
calculate the number of duplicates.

We use list columns in the results, with typically short lists,
which in general can be printed nicely. As tibbles hide elements of
list columns, we return a \code{data.frame}.
}
