% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_project_packages}
\alias{find_project_packages}
\alias{needed_project_packages}
\title{List packages used in a project}
\usage{
find_project_packages(path = ".")

needed_project_packages(path = ".")
}
\arguments{
\item{path}{Relative, or absolute path to a project.}
}
\value{
A character vector with the names of packges loaded.
}
\description{
Find packages loaded in \code{\link{library}}, or \code{\link{require}} calls
in a non-package project.
}
\details{
These functions look for \code{\link{library}}, or \code{\link{require}}
calls in \code{*.R}, \code{*.Rmd}, \code{*.Rnw}, \code{*.Rtex}, and
\code{*.Rrst} files.  \code{needed_project_packages} list all pacakges that
are not installed.
}
\section{Functions}{
\itemize{
\item \code{needed_project_packages}: Only list packages that are not installed.
}}

