% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{get_chunk_paths}
\alias{get_chunk_paths}
\title{Create Hive-style partition paths}
\usage{
get_chunk_paths(
  dataset_base_name,
  file_nrow,
  chunk_size,
  chunk_col_name = "chunk",
  chunk_file_name = "data.parquet"
)
}
\arguments{
\item{dataset_base_name}{Path of the directory to write the Hive
partitioned Parquet files to.}

\item{file_nrow}{Number of rows in \code{file}. As there is no
reliable and cross-platform way to get the exact number of lines
in a compressed file, this has to be set manually to calculate
the number of chunks and the names of partitions. Use \code{wc}
on a Unix-like system to determine row count (\code{zcat file.gz
| wc -l}, or similar).}

\item{chunk_size}{The number of rows to include in each chunk}

\item{chunk_col_name}{Name of the column indicating partition
numbers in the Hive-style partition structure.}

\item{chunk_file_name}{Name of the individual Parquet files in the
Hive-style partition structure.}
}
\value{
A character vector with the paths to the partitions.
}
\description{
To be used with \code{\link{read_delim_chunked_to_dataset}}.
}
\keyword{internal}
