% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{is_num_chr}
\alias{is_num_chr}
\title{Check if a character vector can be converted to numeric}
\usage{
is_num_chr(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A single logical
}
\description{
Check if a character vector can be converted to numeric
}
