% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{is_whole}
\alias{is_whole}
\title{Check whether a vector has only whole values}
\usage{
is_whole(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A single logical
}
\description{
Check whether a vector has only whole values
}
