% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{mark_to_logical}
\alias{mark_to_logical}
\title{Convert character mark to logical}
\usage{
mark_to_logical(x, mark = "X", na_to_false = FALSE)
}
\arguments{
\item{x}{A character vector.}

\item{mark}{A string representing \code{TRUE}.}

\item{na_to_false}{Convert \code{NA}s to \code{FALSE}?}
}
\value{
A logical vector
}
\description{
A helper function to convert a character vector to logical where a single character
string represents \code{TRUE}.
}
