% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{maybe_int}
\alias{maybe_int}
\title{Check whether a vector can be coerced to integer}
\usage{
maybe_int(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A single logical value
}
\description{
Check whether a vector can be coerced to integer
}
