% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{metadata_parquet}
\alias{metadata_parquet}
\title{Query metadata of a Parquet file}
\usage{
metadata_parquet(file)
}
\arguments{
\item{file}{Path to a Parquet file}
}
\value{
A data frame with two columns:

\describe{
  \item{\code{name}}{Column names}
  \item{\code{type}}{Column types, as returned by the
  \code{\link[arrow:ParquetFileReader]{ParquetFileReader$GetSchema()}}
  method. This does not directly correspond to navtive R data types
  but should be informative nevertheless.}
}
}
\description{
Query column names and column types of an Apache Parquet data file.
}
\examples{
tmp <- tempfile()
arrow::write_parquet(iris, tmp)
metadata_parquet(tmp)

}
\seealso{
\code{\link[arrow:write_parquet]{write_parquet}},
  \code{\link[arrow:read_parquet]{read_parquet}}
}
\author{
András Svraka
}
