% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{parse_date_ymd}
\alias{parse_date_ymd}
\title{Parse date string without century}
\usage{
parse_date_ymd(x, cutoff_2000 = 30L)
}
\arguments{
\item{x}{A character vector of dates to parse.}

\item{cutoff_2000}{integer. For \code{y} format,  two-digit numbers smaller or equal
to \code{cutoff_2000} are parsed as though starting with \code{20}, otherwise parsed
as though starting with \code{19}. Available only for functions relying on
\code{lubridate}s internal parser.}
}
\value{
A vector of class \code{\link[base:Dates]{base::Date()}}
}
\description{
Helper function to convert dates stored as strings containing six
digits in a \code{yymmdd} format using \code{\link[lubridate:ymd]{lubridate::ymd()}}.
}
\details{
\code{cutoff_2000} defaults to 30 which is a convenient value
for my work.
}
