% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pretty_print_date}
\alias{pretty_print_date}
\title{Pretty print date in several languages}
\usage{
pretty_print_date(x, locale = c("en_GB", "en_US", "hu_HU"))
}
\arguments{
\item{x}{a date-time object}

\item{locale}{Language to use.}
}
\description{
Format dates according to the conventions of several languages.
}
\details{
Currently only UK and US English, and Hungarian are
  supported.
}
\examples{
locales <- c("en_GB", "en_US", "hu_HU")
sapply(locales, function(x) pretty_print_date(Sys.Date(), x))
}
