% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_chunked}
\alias{read_chunked}
\title{Helper function to select the appropriate readr function}
\usage{
read_chunked(input_delim, ...)
}
\arguments{
\item{input_delim}{Single character used to separate fields within a record.}

\item{...}{Arguments passed to the \code{\link[readr]{read_delim_chunked}}
functions.}
}
\description{
Select the appropriate \code{\link[readr]{read_delim_chunked}} function.
}
\keyword{internal}
