% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{simplify_chunked_codebook}
\alias{simplify_chunked_codebook}
\title{Simplify chunked codebook}
\usage{
simplify_chunked_codebook(codebook)
}
\arguments{
\item{codebook}{A list of lists from \code{\link{codebook}}.}
}
\description{
Simplify a list in \code{\link{codebook_chunked}} into a list of three
tibbles like the output of \code{\link{codebook}}.
}
\keyword{internal}
