% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsummary.R
\name{unnest_modelsummary_list}
\alias{unnest_modelsummary_list}
\title{Unnest a \code{modelsummary_list} column}
\usage{
unnest_modelsummary_list(data, col = "model", remove = TRUE)
}
\arguments{
\item{data}{A data.frame}

\item{col}{Name of the \code{modelsummary_list} column, a character
vector of length one.}

\item{remove}{If ‘TRUE’, remove input column from output data
frame.}
}
\value{
The original data.frame with two added columns, optionally
the input column removed.
}
\description{
Unnest a nested data frame with a \code{modelsummary_list} column by
adding \code{tidy} and \code{glance} columns.
}
\details{
A \code{modelsummary_list} object can be directly used to print
regression tables but needs to be unnested for further processing,
e.g. for plotting from the "tidy" data.frame.
\code{\link[tidyr:unnest_wider]{tidyr::unnest_wider()}} doesn't work for objects of
this type, and it cannot be extended with custom methods. This
function reimplements unnesting in a simple case.
}
