% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{yesno_to_logical}
\alias{yesno_to_logical}
\title{Convert Yes/No values to logical}
\usage{
yesno_to_logical(x, yn = c("I", "N"), na_to_false = FALSE)
}
\arguments{
\item{x}{A character vector.}

\item{yn}{A character vector of length 2, containing the strings
for "yes" (first element) and "no" (second element). Defaults to
Hungarian shorthand for "igen" and "nem".}

\item{na_to_false}{Convert \code{NA}s to \code{FALSE}?}
}
\value{
A logical vector
}
\description{
A helper function to convert a two-valued character vector to a
logical vector.
}
