% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsummary.R
\name{as.modelsummary_list_custom}
\alias{as.modelsummary_list_custom}
\alias{as.modelsummary_list_custom.default}
\alias{as.modelsummary_list_custom.fixest}
\title{Create a custom \code{modelsummary_list}}
\usage{
as.modelsummary_list_custom(x, ...)

\method{as.modelsummary_list_custom}{default}(x, ...)

\method{as.modelsummary_list_custom}{fixest}(x, conf.int = TRUE, ...)
}
\arguments{
\item{x}{A model object}

\item{...}{Passed on to \code{\link[broom:reexports]{broom::tidy()}} and \code{\link[broom:reexports]{broom::tidy()}}}

\item{conf.int}{Logical indicating whether or not to include a
confidence interval in the tidied output. Defaults to \code{TRUE}.}
}
\value{
A \code{modelsummary_list} object, see \href{https://modelsummary.com/vignettes/modelsummary_extension.html}{\pkg{modelsummary}'s Extension and Customization}
vignette.
}
\description{
Create a customized \code{modelsummary_list} from a model object.
}
\details{
Motivation for this generic comes from the need to save
\pkg{fixest} models. Saving models even with \code{lean = TRUE} can
result in hard to serialize objects, as \pkg{fixest} stores calls
and environments which are difficult to serialise.

A \code{modelsummary_list} object made with \code{modelsummary(models, output = "modelsummary_list")} does most of what we want, except for
adding confindence intervals. \pkg{modelsummary} can add confidence
intervals to tables but want to store it here, so that it is
available for any other use, including plotting. Therefore we use
\code{\link[broom:reexports]{broom::tidy()}} to obtain information on parameters. For the
"glance" part we use \code{\link[broom:reexports]{broom::tidy()}} in the default method but we
use data obtained from \pkg{modelsummary}'s custom internal glance
implementation, as it provides useful additions for \pkg{fixest}
models (like adding fixed effects, which we also extend via
\pkg{modelsummary}'s customization interface, see
\code{\link[=glance_custom.fixest]{glance_custom.fixest()}}, or proper information on the types of
standard errors).

The drawback of this approach is that we can't use some
\pkg{modelsummary} features, like \href{https://modelsummary.com/vignettes/modelsummary.html#reference-categories}{reference categories}
and some computations are run twice. This shouldn't be a major
issue for most models and the extra features can be easily
replicated by modifying the \code{modelsummary_list} object.
}
