% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{chr_values}
\alias{chr_values}
\title{Helper function to list top values of a character vector}
\usage{
chr_values(x, n = 5)
}
\arguments{
\item{x}{A character vector}

\item{n}{How many values should be reported}
}
\description{
List the first \code{n} unique values that cannot be coerced into
numeric of a character vector.
}
\keyword{internal}
