% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{Create a codebook from a data frame}
\usage{
codebook(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A tibble from a custom wide format \code{\link[skimr]{skim}} output.  Unlike
\code{\link[skimr]{skim}}, histograms are not generated but there are three
additional columns: \code{\link[sjmisc:is_float]{is_whole}},
\code{\link[sjmisc:is_num_fac]{is_num_chr}}, and \code{chr_values}, a list
 column with the first 10 unique character values.
}
\description{
Create a codebook of a data frame using \code{\link[skimr]{skim}}.
}
\examples{
codebook(mtcars)
}
\seealso{
\code{\link{codebook_chunked}} for files that are too big to be read
  into memory at once.
}
