% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{Create a codebook from a data frame}
\usage{
codebook(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A list of three tibbles.

\describe{
  \item{chr_cols}{A tibble of character columns from \code{df} in rows with a
    vector of the first 10 unique values of the original variable.  It allows
    you to inspect character columns that can actually be encoded as numeric
    columns. Especially useful when the rows higher than \code{guess_max}
    from \code{\link[readr]{read_delim}} contain \code{NA}s.}
  \item{skims}{A tibble with the results from a customized
    \code{\link[skimr]{skim}} output in a wide format}
  \item{problems}{Results from \code{\link[readr]{problems}}}
}
}
\description{
Create a codebook of a data frame using \code{\link[skimr]{skim}}.
}
\examples{
codebook(mtcars)
}
\seealso{
\code{\link{codebook_chunked}} for files that are too big to be read
  into memory at once.
}
