% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook_chunked}
\alias{codebook_chunked}
\title{Create codebook by chunks}
\usage{
codebook_chunked(file, delim, chunk_size, process_function = NULL,
  trim_ws = TRUE, ...)
}
\arguments{
\item{file}{Path to a file.  Passed to
\code{\link[readr]{read_delim_chunked}}, see further details there.}

\item{delim}{Single character used to separate fields within a record.}

\item{chunk_size}{The number of rows to include in each chunk.  See
\code{\link[readr]{read_delim_chunked}}}

\item{process_function}{A function that takes a data frame as its input and
processes this data frame.  \code{process_function} is applied to each
chunk before creating the codebook for that chunk.  There are no
restrictions what this function can be.  When left as \code{NULL}, no
processing is done.}

\item{trim_ws}{Defaults to \code{TRUE} to mimic \code{\link[readr]{read_csv}}
behaviour.}

\item{...}{Additional arguments passed to
\code{\link[readr]{read_delim_chunked}}}
}
\value{
A tibble like from \code{\link{codebook}} with number of variables
\eqn{\times}{*} number of chunks rows and an additional column named
\code{chunk} containing the chunk number.  Parsing problmes by
\code{\link[readr]{read_delim_chunked}} are stored in an attribute of the
 tibble, which can be accessed by \code{\link[readr]{problems}}.
}
\description{
\code{codebook_chunked} uses \code{codebook} to create codebooks for chunks
of a delimited file, as read by \code{\link[readr]{read_delim_chunked}}.
}
\examples{
codebook_chunked(
  file = readr::readr_example("mtcars.csv"),
  delim = ",",
  chunk_size = 10
)
}
