% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_project_packages}
\alias{find_project_packages}
\title{List packages used in a project}
\usage{
find_project_packages(path = ".", not_installed = FALSE)
}
\arguments{
\item{path}{Relative, or absolute path to a project.}

\item{not_installed}{Only list packages that are not installed.}
}
\value{
A character vector with the names of packges loaded.
}
\description{
Find packages loaded in \code{\link{library}}, or \code{\link{require}} calls
in a non-package project.
}
\details{
This function looks for \code{\link{library}}, or \code{\link{require}}
calls in \code{*.R}, \code{*.Rmd}, \code{*.Rnw}, \code{*.Rtex}, and
\code{*.Rrst} files.
}
