% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixest.R
\name{glance_custom.fixest}
\alias{glance_custom.fixest}
\title{Custom glances for fixest modelsummaries}
\usage{
\method{glance_custom}{fixest}(x, ...)
}
\arguments{
\item{x}{model or other R object to convert to single-row data frame}

\item{...}{ellipsis}
}
\description{
Add model family to glances in fixest modelsummaries using the
\code{\link[modelsummary:glance_custom]{modelsummary::glance_custom()}} customization interface.
}
