% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{parse_date_ymd}
\alias{parse_date_ymd}
\title{Parse date string without century}
\usage{
parse_date_ymd(x, orders = "\%y\%m\%d", cutoff_2000 = 30L)
}
\arguments{
\item{x}{a character or numeric vector of dates}

\item{orders}{a character vector of date-time formats. Each order string is
series of formatting characters as listed \code{\link[base:strptime]{base::strptime()}} but might not
include the "\%" prefix, for example "ymd" will match all the possible
dates in year, month, day order.  Formatting orders might include
arbitrary separators. These are discarded.  See details for implemented
formats.}

\item{cutoff_2000}{integer. For \code{y} format,  two-digit numbers smaller or equal to
\code{cutoff_2000} are parsed as 20th's century, 19th's otherwise. Available only
for functions relying on \code{lubridate}s internal parser.}
}
\value{
A character vector of class \code{\link[base]{Date}}
}
\description{
Helper function to convert dates stored as strings containing six
digits using \code{\link[lubridate]{parse_date_time2}}.
}
\details{
\code{cutoff_2000} defaults to 30 which is a convenient
  value for my work. Unlike
  \code{\link[lubridate]{parse_date_time2}}, this function returns
  a date.
}
